requirements.txt
pyzbar==0.1.9
Pillow==10.0.0
reportlab==4.0.4
pywin32==306
numpy==1.24.3
opencv-python==4.8.0
qrcode==7.4.2
flask==2.3.3 